FUZZ-231 Eliminated exhaust function to simplify BasePrimitve interface.
exhaust() was used exactly once. Required refactor in Session(). It did two things:
 1. Return how many cases were being skipped
 2. Cause mutate to return False, meaning that the next fuzz mutation requested by Session() would return false. Get it? this is a big, crazy, break statement. Unclean! Unclean code!

The fix for 1. was to calculate the value using other public attributes. The fix for 2. was to move that statefulness within Session(). Not super clean, but better than the crazy coupling we had before.

Verfied by running tests (designed to fail and hit the skip count) before and after refactor. Only differences on received time stamps were seen.