Cleanup: Size, BitFiled as per review.
Note: This commit also removes a tiny bit of functionality in which
BitField could receive a list for its value argument instead of just a
numbeer. This was used nowhere as far as I could tell, and was implemented
in a confusing spot. I don't know with certainty if it was working. The
right way to do this is to provide another optional argument to either
supplement or supplant the normal fuzz mutations. This should be a method on
all block types.