Make sure to render primitives as binary strings
The fuzzing framework expects the result of a primitive.render() to be a
bytes-like string.  Under Python 2, the str objects being returned
worked just fine.  For Python 3 compatibility, such strings must be
encoded into a `bytes` object.

For the numeric primitives, this was not a problem, as their render code
(in the BitField class) was using native binary conversions already.
This patch serves to catch unicode strings which slip through via the
string-like primitives (String, Delim, Static, RandomData, from_file).

Signed-off-by: Malf Furious <m@lfurio.us>