Don't encode values in String._render()
Per the previous commit, this encoding fixup is centralized in
BasePrimitive.render().  So it can be removed from this function.

The issue expressed in the comment I've removed _is still a problem_,
but it doesn't make sense to leave it in this spot.  It actually relates
a bit to the code added in the previous commit.

I originally found this comment hard to comprehend.  I think a better
way of saying it, recontextualizing it within Python 2 and 3, may be
thus:

        Currently, the project has a big bias towards "ascii" string
        encoding.  This can be a problem when we need to treat various
        strings as binary (such as for the purposes of rendering a
        primitive).  To me more flexible, a default of some unicode
        encoding may be more desirable.  Prehaps, utf_8.

Signed-off-by: Malf Furious <m@lfurio.us>