Handle TIME_WAIT case for servers
Similar to clients calling connect(), a server can fail in bind() if
there is a previous socket for the same host/port in TIME_WAIT state.
Even SO_REUSEADDR will not completely solve the problem, as connecting
to the same client will still fail.

Raise a BoofuzzOutOfAvailableSockets exeption similar to the client
connect() code, when a bind() call fails with EADDRINUSE. This allows
the session code to retry the connection long enough for TIME_WAIT to
resolve.