add BoofuzzFailure to interrupt test case execution on failure... (#415)
...and fix a string decoding bug in the Unix procmon.

This PR adds `BoofuzzFailure`, an exception that can be thrown when failures are detected to interrupt test case execution and skip to the next test case. This is especially important when the behavior of some messages in the protocol depends on the results from a previous received message. Without a valid response to message 1, message 2 may be poorly defined, hence the need to abort the test case.

Usage: Instead of calling `log_fail()` on the fuzz logger, raise `BoofuzzFailure` with your failure message. The failure will be logged and the test case terminated.

Can be used in protocol definitions, especially in callback methods. Also used internally by boofuzz.