Local procmon (#483)
Local process monitor so as to run targets without a separate procmon command.

The implementation suggests some refactors:

1. Use local procmon as the main procmon implementation. Rewrite the ProcessMonitorPedrpcServer to just have-a ProcessMonitorLocal.
2. Clarify the differences between the server and client side process monitors. I think the method args can be identical -- right now they're mostly identical with subtle differences.
3. Have the local procmon, the client procmon, and the server procmon all inherit from the procmon base class.
However it might be best to merge as is for the purposes of moving forward.

Also slipped in a change for loggers: If you pass no loggers, the Session still creates a default for you. If you pass an empty list however, you're saying you want no loggers, and the console output will be quiet.