Handle TLS handshake timeouts in SSLSocketConnection (#712)
SSLSocketConnection prior to this commit would hang identifitly if a
timeout during the TLS handshake occurs. This can happen if a target
accepts a TCP connection but does not send any TLS handshake messages,
e.g. because of a bug triggered by the fuzzing.

This commit makes boofuzz observe timeouts during the TLS handshake. TLS
handshake timeouts are handled in the same way as timeouts during TCP
connection setup.

Co-authored-by: Lothar Braun <lothar.braun@tha.de>